Moved components/GroupSearch into pages/community/search/index and renamed GroupSearchFilters to be a first-class component. Deleted old debug thing. Fixed up loading/empty state design on Group Search results.
