--- components/GroupSearch/index.js
@@ -1,3 +1,0 @@
-import GroupSearch from './GroupSearch';
-
-export default GroupSearch;
