+++ components/GroupSearchFilters/index.js
@@ -1,0 +1,3 @@
+import GroupSearchFilters from './GroupSearchFilters';
+
+export default GroupSearchFilters;
