+++ pages/community/Community.styles.js
@@ -13,7 +13,7 @@ const Hero = styled.div`
       rgba(246, 246, 246, 0.75) 50%,
       rgba(246, 246, 246, 1) 80%
     ),
-    url(https://source.unsplash.com/random/307x200);
+    url(https://images.unsplash.com/photo-1501785888041-af3ef285b470?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=4050&q=80);
   background-size: cover;
   align-items: center;
   display: flex;
